refactor(ui): extract rendering logic into a dedicated module
This commit separates all rendering and DOM manipulation logic from the core application file into a new `ui/renderer.js` module.

- All functions responsible for generating HTML (`render`, `renderTransactionList`, `renderPersons`, `renderStats`, `renderChart`) now reside in `ui/renderer.js`.
- `core/main.js` is now simpler, focusing on data handling and application setup. It calls the `render` function from the new module.
- Dependencies required by the renderer (like `calculateBalance` and action handlers) are passed in via an `initRenderer` function to maintain clear data flow.
- The `onclick` attributes in the person list have been replaced with event listeners, improving encapsulation.

This change significantly improves the separation of concerns between application logic and its presentation.
